% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Mean Square Error of SBLF model}
\usage{
mse(fit)
}
\arguments{
\item{fit}{SBLF fit object, save SBLF() output for this parameter}
}
\description{
Compute and print the mean square error (MSE) from the training data
and mean square prediction error (MSPE) from the testing data
}
