\name{SBLF-package}
\alias{SBLF-package}
\alias{SBLF}
\docType{package}
\title{
\packageTitle{SBLF}
}
\description{
\packageDescription{SBLF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SBLF}
\packageIndices{SBLF}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SBLF}

Maintainer: \packageMaintainer{SBLF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
