% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voxel_loc}
\alias{voxel_loc}
\title{Matrix of coordinates of voxel locations.}
\format{
An object of class \code{matrix} with 315 rows and 3 columns.
}
\usage{
voxel_loc
}
\description{
The number of rows should match the number of voxels. In
this example, there are 315 voxels.
}
\keyword{datasets}
