% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xtest}
\alias{xtest}
\title{Randomly selected subset of real imaging predictor data for testing}
\format{
An object of class \code{matrix} with 10 rows and 10080 columns.
}
\usage{
xtest
}
\description{
This matrix contains a randomly selected subset of real imaging predictor
data for testing. The data set contains 32 imaging predictors
and 315 voxels per image. \verb{32 * 315 = 10,000} columns, and 10 rows.
}
\keyword{datasets}
