% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xtrain}
\alias{xtrain}
\title{This matrix contains a randomly selected subset
of real imaging predictor data for training.}
\format{
An object of class \code{matrix} with 50 rows and 10080 columns.
}
\usage{
xtrain
}
\description{
The data set contains 32 imaging predictors
and 315 voxels per image. \verb{32 * 315 = 10,000} columns,
and 50 rows.
}
\keyword{datasets}
