% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBLF.R
\name{SBLF}
\alias{SBLF}
\title{Fit Spatial Bayesian Latent Factor Models}
\usage{
SBLF(xtrain, xtest, ztrain, ztest, voxel_loc)
}
\arguments{
\item{xtrain}{training data set of imaging predictors. Sample size is \code{ntrain * (P * image_len)}}

\item{xtest}{test data set of imaging predictors. sample size is \code{(ntotal-ntrain) * (P * image_len)}}

\item{ztrain}{training data set of imaging outcomes. Sample size is \code{(ntrain * image_len)}}

\item{ztest}{test data set of imaging outcomes. Sample size is \code{((ntotal-ntrain) * image_len)}}

\item{voxel_loc}{matrix of voxel coordinates.}
}
\value{
A list which contains
}
\description{
Use this function to fit spatial Bayesian latent factor models given
the prepared input data sets. Please also comment on the important
output the user should be expecting.
}
\details{
The parameters prefixed with "x" are imaging predictor data, and
the parameters prefixed with "z" are imaging outcomes.

Specify imprtant details to the user about the use
of this function. Please comment.

Data sets should be entered an unnamed matrices (see examples for illustration).
The data matrices should not have rownames or column names.
}
\examples{
\donttest{

}

}
\references{
Please add your paper reference.
}
